いい感じにViewモジュールを分割した
