+++ src/js/Components/Models/Config.js
@@ -3,7 +3,14 @@ import {Model} from 'chomex';
 export default class Config extends Model {}
 Config.default = {
   'winconfigs': {
-    value: []
+    value: [
+      {
+        id: '小型',
+        alias: '小型',
+        size: {width: 400, height: 240},
+        decoration: 'EXTRACT'
+      }
+    ]
   },
   'closeconfirm': {
     value: 'おっぱああああい!!'
